requirements.txt
fastapi
uvicorn
pytesseract
opencv-python
Pillow
python-multipart
spacy
